requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
black==26.3.1
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
gunicorn==25.2.0
iniconfig==2.3.0
isort==8.0.1
kombu==5.6.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
pytokens==0.4.1
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
